Fix protection of manage_delObjects with "Delete objects".
In certain situations, when only having the permission to delete the
content ("Delete portal content") but not the permission to delete from the
the parent container ("Delete objects" on the parent) deleting was possible
even though it shouldn't have been.

This should not be that problematic since the user alyways needed the
"Delete portal content" permission on the item.