Synchronize delete permissions.
Let roles with "Delete objects" also have "Delete portal content" on site root.
This is important in order to avoid errors such as seeing a delete action but
getting an error when using it. It does only affect the site root settings and
not any workflows.

--

The idea is that all roles with "Delete objects" permission should
also have the "Delete portal content" permission in order to
actually be able to delete things.
This principle applies to the site root settings and may be changed
by workflows, depending on the use case.